Support the ability to switch between capture and generation modes (#18)
Support the ability to switch between capture and generation modes in real time without restarting the modules.

In the situation of switching from capture mode to PWM generation mode, a one-time delay of about 2-3 seconds is possible, after this delay, the response becomes normal again (we need to find out why).